Fixes nperpage and pagination bug
