+++ src/lib/list/pagination-utils.ts
@@ -84,3 +84,23 @@ export const getPageTiles = (idx: number, nPage: number): any[] => {
 
   return arr;
 };
+
+export const paginationBoundaries = (
+  idx: number,
+  nPerPage: number
+): { start: number; end: number } => {
+  const start = (idx - 1) * nPerPage;
+  const end = idx * nPerPage;
+
+  return { start, end };
+};
+
+export const withPagination = <A>(
+  data: A[],
+  idx: number,
+  nPerPage: number
+): A[] => {
+  const { start, end } = paginationBoundaries(idx, nPerPage);
+
+  return data.slice(start, end);
+};
